fix: plug protocol crashes, LPD wiring, type safety, infrastructure smells
Protocol bugs:
- TrackerServer/UdpTrackerServer: replace Split('.') IP parsing with IPAddress.TryParse + GetAddressBytes(); malformed IPs skipped not crash
- TrackerServer: compact peers now use BString(byte[]) not Latin1 string encoding; binary wire format correct
- LocalPeerDiscovery: wire parsed infoHash+port into IPeerDiscoveryService.AddPeers; LPD peer discovery was completely non-functional
- PeerConnection: byte shifts cast to uint before shift then int; add length<0 guard
- PeerServer: same unsigned-shift fix for index/begin/length in piece request handler
- MseHandshake: ReadUint16 returns ushort not int

Infrastructure:
- Startup: Uri.TryCreate for Origin header; UriFormatException no longer crashable by client
- AssemblyLoader: narrow catch to FileNotFoundException/BadImageFormatException/FileLoadException; log at Warn
- BuildInfo: narrow catch to IOException/FormatException; log at Warn
- StringExtensions: remove dead IsNullOrWhiteSpace/IsNotNullOrWhiteSpace extension methods
- RestController: remove wasted validator allocations in base constructor; properties default null

Frontend:
- Statistics: log tracker URL parse failures instead of silently swallowing

Tests:
- LocalPeerDiscoveryTest: mock IPeerDiscoveryService; assert AddPeers called with correct infoHash+port
- PeerConnectionTest: new test for length<0 rejection on high-bit-set length bytes
- PeerServerTest: 3 new tests for high-bit-set index/begin/length rejection

+++ src/NzbDrone.Core/Peers/Encryption/MseHandshake.cs
@@ -415,9 +415,9 @@ public class MseHandshake
             ((uint)bytes[2] << 8) | bytes[3];
     }
 
-    private static int ReadUint16(byte[] bytes)
+    private static ushort ReadUint16(byte[] bytes)
     {
-        return (bytes[0] << 8) | bytes[1];
+        return (ushort)((bytes[0] << 8) | bytes[1]);
     }
 
     private static byte[] GeneratePadding()
